requirements.txt
anthropic
apscheduler
python-dotenv